Rename redisdb to redis
